visionary-art-ai-service/requirements.txt
blendmodes
accelerate
basicsr
fonts
font-roboto
gfpgan
gradio==3.23
invisible-watermark
numpy
omegaconf
opencv-contrib-python
requests
piexif
Pillow
pytorch_lightning==1.7.7
realesrgan
scikit-image>=0.19
timm==0.4.12
transformers==4.25.1
torch
einops
jsonmerge
clean-fid
resize-right
torchdiffeq
kornia
lark
inflection
GitPython
torchsde
safetensors
psutil
rich
